requirements.txt
python-telegram-bot
aiohttp
pyyaml
telegram